# net/http: let Transport request body writes use sendfile